--- CLI/src/main/java/ru/spbau/sd/cli/IOStream.java
@@ -1,3 +1,0 @@
-package ru.spbau.sd.cli;
-
-interface IOStream extends InputStream, OutputStream {}
