+++ CLI/src/main/java/ru/spbau/sd/cli/interpreter/ExecutionResult.java
@@ -1,4 +1,4 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.interpreter;
 
 public enum ExecutionResult {
     OK, Error, Finish
