+++ CLI/src/main/java/ru/spbau/sd/cli/io/IOStream.java
@@ -1,0 +1,3 @@
+package ru.spbau.sd.cli.io;
+
+public interface IOStream extends InputStream, OutputStream {}
