+++ CLI/src/main/java/ru/spbau/sd/cli/io/OutputStream.java
@@ -1,4 +1,4 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.io;
 
 public interface OutputStream {
     void write(String str);
