+++ CLI/src/test/java/ru/spbau/sd/cli/commands/BuiltInCommandTest.java
@@ -1,10 +1,13 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.commands;
 
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.mockito.Mockito;
+import ru.spbau.sd.cli.commands.BuiltinCommand;
+import ru.spbau.sd.cli.interpreter.ExecutionResult;
+import ru.spbau.sd.cli.io.InputStream;
 
 import java.io.*;
 import java.nio.file.Paths;
@@ -22,7 +25,7 @@ public class BuiltInCommandTest {
 
     @Test
     public void echoArgTest() {
-        OutputStream outputStream = Mockito.mock(OutputStream.class);
+        ru.spbau.sd.cli.io.OutputStream outputStream = Mockito.mock(ru.spbau.sd.cli.io.OutputStream.class);
         BuiltinCommand.echo.run(Collections.singletonList(content),
                 null, outputStream);
         Mockito.verify(outputStream).write(Mockito.eq(content));
@@ -30,9 +33,9 @@ public class BuiltInCommandTest {
 
     @Test
     public void echoStreamTest() {
-        InputStream inputStream = Mockito.mock(InputStream.class);
+        ru.spbau.sd.cli.io.InputStream inputStream = Mockito.mock(InputStream.class);
         Mockito.when(inputStream.read()).thenReturn(content);
-        OutputStream outputStream = Mockito.mock(OutputStream.class);
+        ru.spbau.sd.cli.io.OutputStream outputStream = Mockito.mock(ru.spbau.sd.cli.io.OutputStream.class);
         BuiltinCommand.echo.run(Collections.emptyList(), inputStream, outputStream);
         Mockito.verify(outputStream).write(Mockito.eq(content));
     }
@@ -43,7 +46,7 @@ public class BuiltInCommandTest {
         Writer writer = new FileWriter(file);
         writer.write(content);
         writer.close();
-        OutputStream outputStream = Mockito.mock(OutputStream.class);
+        ru.spbau.sd.cli.io.OutputStream outputStream = Mockito.mock(ru.spbau.sd.cli.io.OutputStream.class);
         BuiltinCommand.cat.run(Collections.singletonList(file.toPath().toString()),
                 null, outputStream);
         Mockito.verify(outputStream).write(Mockito.eq(content + '\n'));
@@ -59,7 +62,7 @@ public class BuiltInCommandTest {
     @Test
     public void pwdTest() {
         String dir = Paths.get(".").toAbsolutePath().normalize().toString();
-        OutputStream outputStream = Mockito.mock(OutputStream.class);
+        ru.spbau.sd.cli.io.OutputStream outputStream = Mockito.mock(ru.spbau.sd.cli.io.OutputStream.class);
         BuiltinCommand.pwd.run(Collections.emptyList(), null, outputStream);
         Mockito.verify(outputStream).write(dir);
     }
@@ -71,7 +74,7 @@ public class BuiltInCommandTest {
         writer.write(content + '\n');
         writer.write(content + '\n');
         writer.close();
-        OutputStream outputStream = Mockito.mock(OutputStream.class);
+        ru.spbau.sd.cli.io.OutputStream outputStream = Mockito.mock(ru.spbau.sd.cli.io.OutputStream.class);
         BuiltinCommand.wc.run(Collections.singletonList(file.toPath().toString()),
                 null, outputStream);
         List<String> expected = Stream.of(2, 4, 2 * (content.length() + 1),
