+++ CLI/src/test/java/ru/spbau/sd/cli/interpreter/ASTElementTest.java
@@ -1,7 +1,10 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.interpreter;
 
 import org.junit.Test;
 import org.mockito.Mockito;
+import ru.spbau.sd.cli.commands.Command;
+import ru.spbau.sd.cli.interpreter.ASTElement;
+import ru.spbau.sd.cli.interpreter.ExecutionResult;
 
 import java.util.Collections;
