+++ CLI/src/test/java/ru/spbau/sd/cli/interpreter/InterpreterSessionTest.java
@@ -1,7 +1,8 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.interpreter;
 
 import org.junit.Assert;
 import org.junit.Test;
+import ru.spbau.sd.cli.interpreter.InterpreterSession;
 
 public class InterpreterSessionTest {
     private final String content = "content";
