+++ CLI/src/test/java/ru/spbau/sd/cli/interpreter/ParserTest.java
@@ -1,8 +1,15 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.interpreter;
 
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
+import ru.spbau.sd.cli.commands.AssignmentCommand;
+import ru.spbau.sd.cli.commands.BuiltinCommand;
+import ru.spbau.sd.cli.commands.Command;
+import ru.spbau.sd.cli.commands.ExternalCommand;
+import ru.spbau.sd.cli.interpreter.ASTElement;
+import ru.spbau.sd.cli.interpreter.Environment;
+import ru.spbau.sd.cli.interpreter.Parser;
 
 import java.util.Arrays;
