+++ CLI/src/test/java/ru/spbau/sd/cli/interpreter/SimpleEnvironmentTest.java
@@ -1,7 +1,9 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.interpreter;
 
 import org.junit.Assert;
 import org.junit.Test;
+import ru.spbau.sd.cli.interpreter.Environment;
+import ru.spbau.sd.cli.interpreter.SimpleEnvironment;
 
 public class SimpleEnvironmentTest {
     @Test
