+++ CLI/src/test/java/ru/spbau/sd/cli/io/SimpleStreamTest.java
@@ -1,7 +1,8 @@
-package ru.spbau.sd.cli;
+package ru.spbau.sd.cli.io;
 
 import org.junit.Assert;
 import org.junit.Test;
+import ru.spbau.sd.cli.io.SimpleStream;
 
 public class SimpleStreamTest {
     private final String testStr1 = "foo";
