+++ settings.gradle
@@ -8,4 +8,20 @@
  */
 
 rootProject.name = 'repo'
+//0
 include 'CLI'
+//1
+//2
+//3
+//4
+//5
+//6
+//7
+//8
+//9
+//10
+//11
+//12
+//13
+//14
+//15
